+++ ProyectoDWES/src/main/java/dao/PersonaDao.java
@@ -56,7 +56,7 @@ public class PersonaDao {
 			ps.setString(1, email);
 			rs = ps.executeQuery();
 			if (rs.next())
-				return new Persona(rs.getInt(1),rs.getString(2), rs.getString(3));
+				return new Persona(rs.getLong(1),rs.getString(2), rs.getString(3));
 
 		} catch (SQLException e) {
 			System.out.println("Error al consultar por email " + e.getMessage());
